pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["quanganhats <anhtdq@ats.com.vn>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.28.1"
fastapi = "^0.110.0"
pymongo = "^4.6.2"
motor = "^3.3.2"
pydantic-settings = "^2.2.1"
pydantic = "^2.6.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
beanie = "^1.25.0"
pandas = "^2.2.2"
python-multipart = "^0.0.9"
openpyxl = "^3.1.5"

[tool.poetry.scripts]
start-app = "app.main:app"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
